Fix UCT finding draw as best move when winning
We now clear the TTable for the entries in our history.  It's not
perfect, as we could still hit an entry further down in the tree that is
a repetition.  But this is the fastest way to fix this issue.

The issue was the Network seems to like rep-draw moves, and so then
those nodes get explored a lot.  Then, it accumulates enough visits so
it gets considered the best, even though all the other moves are better
(continue the game and don't draw it immediately).