Update UCTSearch.cpp
To correct the broken conversion from winrate to centipawns. The current conversion is far too conservative and never reaches values extreme enough to allow GUIs to adjudicate win or loss. This is a major problem when trying to set up matches with long time control against other engines. The new suggested value of 305.66 in the conversion function equates a 95% winrate with a 900 centipawn advantage. 

This value is taken from the DeepMind AlphaZero paper:
"Settings were chosen to correspond with computer chess tournament conditions: each player was allowed 1 minute per move, resignation was enabled for all players (-900 centipawns for 10consecutive moves for Stockfish and Elmo, 5% winrate for AlphaZero)."