Update TimeMan.cpp
Seems to be a good compromise between 100 and 178. Everyone sees in TCEC that we play too fast. I think any improvements should be compared to this setting. One thing we haven't tried is using time.maximum every time with standard slow mover, and relying on treereuse and timemanage to save 79% time...