Prolong cfg_lagbuffer to 100ms
Flagging was observed at very fast time controls on Google Compute 8-CPU instance e.g. 40 moves in 20 seconds. Increasing the lag buffer to 100ms has fixed this. It is better to solve this by increasing cfg_lagbuffer than by requiring changes to slowmover by the user. The new setting seems ok down to time-controls as fast as 40 moves in 2 seconds (!)

I suspect this was only manifesting itself when ```else if (elapsed_millis < 1000 || playouts < 100) {
        // Until we reach 1 second or 100 playouts playout_rate
        // is not reliable, so just return max.
        return MAXINT_DIV2;
```
was true and no pruning was occurring.