Improve the mobile view
Make the mobile view look like a scaled-down version of the desktop view, rather than a broken mess.

This is very far from an ideal mobile view, but it’s an undeniable improvement.

I incidentally upgraded to the latest versions of Vega and Vega-Lite, used to draw the progress chart on the homepage, because it seemed like wasted effort to make it work with an old version when presumably at some point we’ll upgrade.